requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.3
Pillow==10.1.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
matplotlib==3.8.2
werkzeug==3.0.1
python-dotenv==1.0.0
pytorch-grad-cam>=1.4.0
